api/KEVIN_ai-cognitive-assessment/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
av==16.0.1
cartesia
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.10.6
ctranslate2==4.6.1
exceptiongroup==1.3.0
fastapi==0.116.1
faster-whisper==1.2.1
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.23.2
packaging==25.0
pillow==11.0.0
pluggy==1.6.0
protobuf==6.33.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv
PyYAML==6.0.3
regex==2025.8.29
requests==2.32.5
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.3
sympy==1.13.1
tiktoken==0.11.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typer-slim==0.20.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
